Issue with the black screen when the game starts
In our custom HTML shell, the default Godot loading splash and progress indicators aren't automatically included because we're overriding the built-in template. This leads to the black screen while the browser downloads and initializes the .wasm and .pck files (which can take a few seconds depending on file sizes and connection speed). To fix this, we'll modify custom_shell.html to manually add a loading screen that displays the boot splash image (the Godot logo by default, unless you've changed it in Project Settings > Application > Boot Splash > Image). We'll also hook into the engine's startup promise to hide the loading screen once everything is ready.